refactor: consolidate integration insert tests
